requirements.txt
fastapi
uvicorn
groq
python-dotenv
pydantic
openai
sqlalchemy
psycopg2 
langchain-mcp-adapters
mcp
aiosmtplib
email-validator
